refactor(de): extract modules from monolithic main.rs
Split the 1275-line single-file DE crate into focused modules, mirroring
the pattern applied to SCS in 1baf26a:

- util.rs: resource/dependency helpers, QID parsing, diagnostic severity,
  RTQ enqueue helper
- backoff.rs: exponential backoff constants and duration computation
- finder.rs: package finder composition (user → cross-repo → CDB → std)
- worker/mod.rs: Worker struct, run_loop, work, and state handlers
  (run_desired, run_lingering, run_undesired, destroy_untouched_resources,
  publish_diagnostics, build_cross_repo_finder)
- worker/eval.rs: compile_and_evaluate and EvalOutcome

main.rs now contains only the CLI, main(), process() (worker scheduling),
and deployment_owned_by_worker (sharding hash). No behavioral changes;
purely structural.
